Module B00_cli.Arg

Command line arguments.

Specifying output format

type output_format = [

| `Normal

| `Short

| `Long

]

The type for specifying output format.

val output_format : ?docs:string -> ?short_opts:string list -> ?long_opts:string list -> unit -> output_format Cmdliner.Term.t

output_format ~short_opts ~long_opts () are mutually exclusive options to specify short and long output format, without options this is `Normal. short_opts defaults to ["s"; "short"] and long_opts default to ["l"; "long"]. docs is the manual section in which options are documented, defaults to s_output_format_options.