Module B00_github

GitHub interaction.

GitHub authentication

module Auth : sig ... end

GitHub API requests

type content_type = string

The type for content types.

type v3_body = [

| `Json of B00_serialk_json.Jsong.t

| `Other of content_type * string

| `Empty

]

The type for GitHub V3 request bodies. Either JSON or something else tagged with its content type or nothing.

val req_json_v3 : ?headers:B00_http.Http.headers -> B00_http.Httpr.t -> Auth.t -> path:string -> B00_http.Http.meth -> v3_body -> (B00_serialk_json.Json.t, string) result

req_json_v3 auth path m ~headers body performs the request for json on path using method m, additional headers headers, body body and authentication auth.

val query_v4 : B00_http.Httpr.t -> Auth.t -> string -> (B00_serialk_json.Json.t, string) result

query_v4 auth q performs the the GitHub GraphQL V4 query q using authentication auth.

Higher-level interface

module Repo : sig ... end

GitHub repositories.

module Issue : sig ... end

Repository issues.

module Release : sig ... end

Repository releases.

module Pages : sig ... end

Publish to GitHub pages.