Module B00_htmlg.At

Element attributes.

Attributes

type name = string

The type for attribute names.

type t

The type for attributes.

val v : name -> string -> t

v n value is an attribute named n with value value.

val true' : name -> t

true' n is v n "". This sets the boolean attribute n to true. The attribute must be omitted to be false.

val int : name -> int -> t

int n i is v n (string_of_int i).

val add_if : bool -> t -> t list -> t list

add_if c att atts is att :: atts if c is true and atts otherwise.

val add_if_some : name -> string option -> t list -> t list

add_if_some n o atts is (v n value) :: atts if o is Some value and atts otherwise.

val to_pair : t -> string * string

to_pair at is (n,v) the name and value of the attribute.

Attribute names and constructors

See the MDN HTML attribute reference.

Convention. Whenever an attribute name conflicts with an OCaml keyword we prime it, see for example class'.

module Name : sig ... end

Attribute names.

type 'a cons = 'a -> t

The type for attribute constructors with value of type 'a.

val accesskey : string cons

accesskey

val autofocus : t

autofocus

val charset : string cons

charset

val checked : t

checked

val class' : string cons

class

val content : string cons

content

val contenteditable : bool cons

contenteditable

val cols : int cons

cols

val defer : t

defer

val disabled : t

disabled

val dir : string cons

dir

val draggable : bool cons

draggable

val for' : string cons

for'

val height : int cons

height

val href : string cons

href

val hidden : t

hidden

val id : string cons

id

val lang : string cons

lang

val media : string cons

media

val name : string cons

name

val placeholder : string cons

placeholder

val rel : string cons

rel

val required : t

required

val rows : int cons

rows

val spellcheck : string cons

spellcheck

val src : string cons

src

val tabindex : int cons

tabindex

val title : string cons

title

val type' : string cons

type

val value : string cons

value

val wrap : string cons

wrap

val width : int cons

width